backend/requirements.txt

file-utils==0.2.1
google-cloud-aiplatform==1.105.0
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.8
langchain-huggingface==0.3.1
markitdown==0.1.2
Pillow==11.3.0
protobuf==6.31.1
python-dotenv==1.1.1
torch==2.7.1
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-genai
pdfplumber
docx
python-docx
pytesseract
pillow
tesseract
python-multipart
tiktoken>=0.5.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
# Setuptools for local development
setuptools>=67.7.2
-e .
